requirements.txt
aiogram==3.18.0
aiocron==2.1
python-dotenv==1.0.1
tinkoff-investments==0.2.0b109
asyncpg==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
pydantic==2.10.6
alembic==1.14.1
asyncio==3.4.3